+++ src/lib.rs
@@ -20,6 +20,6 @@ pub use {
     },
 };
 
-const ASPECT_RATIO: f32 = 4.0 / 3.0;
+const ASPECT_RATIO: f64 = 4.0 / 3.0;
 pub const IMAGE_WIDTH: i32 = 800;
-pub const IMAGE_HEIGTH: i32 = (IMAGE_WIDTH as f32 / ASPECT_RATIO) as i32;
+pub const IMAGE_HEIGTH: i32 = (IMAGE_WIDTH as f64 / ASPECT_RATIO) as i32;
